reconstruct test hook
