+++ tests/test_hook/logger/test_mflow_hook.py
@@ -1,0 +1,36 @@
+import shutil
+import sys
+from unittest.mock import MagicMock
+
+import pytest
+import torch
+from tests.test_hook.test_utils import _build_demo_runner
+from torch.utils.data import DataLoader
+
+from mmcv.runner.hooks import MlflowLoggerHook
+
+
+@pytest.mark.parametrize('log_model', (True, False))
+def test_mlflow_hook(log_model):
+    sys.modules['mlflow'] = MagicMock()
+    sys.modules['mlflow.pytorch'] = MagicMock()
+
+    runner = _build_demo_runner()
+    loader = DataLoader(torch.ones((5, 2)))
+
+    hook = MlflowLoggerHook(exp_name='test', log_model=log_model)
+    runner.register_hook(hook)
+    runner.run([loader, loader], [('train', 1), ('val', 1)])
+    shutil.rmtree(runner.work_dir)
+
+    hook.mlflow.set_experiment.assert_called_with('test')
+    hook.mlflow.log_metrics.assert_called_with(
+        {
+            'learning_rate': 0.02,
+            'momentum': 0.95
+        }, step=6)
+    if log_model:
+        hook.mlflow_pytorch.log_model.assert_called_with(
+            runner.model, 'models')
+    else:
+        assert not hook.mlflow_pytorch.log_model.called
